Refactoring Login

+++ backend/fitness_centre/src/main/java/com/fitness_centre/service/impl/UserServiceImpl.java
@@ -1,15 +1,19 @@
 package com.fitness_centre.service.impl;
 
-import com.baomidou.mybatisplus.core.conditions.Wrapper;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.support.SFunction;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.fitness_centre.domain.LoginUser;
+import com.fitness_centre.constant.ErrorCode;
+import com.fitness_centre.security.LoginUser;
 import com.fitness_centre.domain.User;
 import com.fitness_centre.dto.GeneralResponseResult;
 import com.fitness_centre.dto.UserLoginRequest;
 import com.fitness_centre.dto.UserRegisterRequest;
+import com.fitness_centre.exception.AuthException;
+import com.fitness_centre.exception.BusinessException;
+import com.fitness_centre.exception.SystemException;
+import com.fitness_centre.exception.ValidationException;
 import com.fitness_centre.mapper.UserMapper;
 import com.fitness_centre.service.MailService;
 import com.fitness_centre.service.UserService;
@@ -23,24 +27,17 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.sql.SQLIntegrityConstraintViolationException;
 import java.time.LocalDateTime;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Function;
 
 /**
  * @author
@@ -69,6 +66,8 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     //验证码失效的时间
     private final static int emailExpireTime = 5;
 
+    private final static int basicInfoExpireTime = 45;
+
     @Autowired
     private RecaptchaValidator recaptchaValidator;
 
@@ -88,30 +87,17 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     @Override
     public GeneralResponseResult login(UserLoginRequest loginRequest) {
 
-
-        // bot check
-        boolean recaptchaPassed = recaptchaValidator.verify(
-                loginRequest.getRecaptchaToken(),
-                threshold,
-                "login"
-        );
-
-        if(!recaptchaPassed){
-            throw new BadCredentialsException("Suspected robot");
-        }
-
         //login check
         UsernamePasswordAuthenticationToken authenticationToken =
                 new UsernamePasswordAuthenticationToken(
                         loginRequest.getEmail(),
                         loginRequest.getPassword());
-        Authentication authentication = null;
-        try {
+        Authentication authentication  = null;
+        try{
             authentication = authenticationManager.authenticate((authenticationToken));
-        } catch (Exception e) {
-            throw new BadCredentialsException("The email address or password is incorrect");
+        } catch (AuthenticationException ex) {
+            throw new AuthException(ErrorCode.FORBIDDEN.getCode(),ex.getMessage());
         }
-
 
 
         LoginUser loginUser = (LoginUser) authentication.getPrincipal();
@@ -123,7 +109,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
         Map<String, String> map = new HashMap<>();
         map.put("token", jwt);
 
-        return new GeneralResponseResult(HttpStatus.OK.value(), "Login Successful", map);
+        return new GeneralResponseResult(ErrorCode.SUCCESS, map);
     }
 
     //登出
@@ -135,75 +121,85 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
         String email = loginUser.getUser().getEmail();
         boolean flag = redisCache.deleteObject("login:" + email);
         if (flag == false){
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,"Failed to exit");
+            throw new AuthException(ErrorCode.CACHE_ERROR.getCode(),"Failed to exit");
         }
-        return new GeneralResponseResult<>(HttpStatus.OK.value(), "Exit successfully");
+        return new GeneralResponseResult<>(ErrorCode.SUCCESS);
     }
 
     //todo 频繁调用
     //发送验证码
     @Override
-    public GeneralResponseResult sendCode(UserRegisterRequest request)  {
+    public GeneralResponseResult basicInfoStore(UserRegisterRequest request)  {
 
-        // 机器人检测
-        boolean recaptchaPassed = recaptchaValidator.verify(
-                request.getRecaptchaToken(),
-                threshold,
-                "register"
-        );
 
-        if(!recaptchaPassed){
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Suspected robot");
-        }
         //验证两次密码是否相同
         if(!request.confirmPasswordValid()){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"The confirmation password must be the same as the password");
+            throw new ValidationException(ErrorCode.PASSWORDS_DO_NOT_MATCH);
         }
         //检查邮箱是否被注册了
         boolean exists = this.lambdaQuery()
                 .eq(User::getEmail,request.getEmail())
                 .exists();
         if(exists){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Email already in use");
+            throw new AuthException(ErrorCode.USER_ALREADY_EXISTS);
         }
 
-        //控制发送的频率,一分钟发一次
-        String sendFreq = "emailSendFreq:" + request.getEmail();
-
-        if(!Objects.isNull(redisCache.getCacheObject(sendFreq))){
-            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS,"You have sent verification code recently, please wait.");
-        }
-        redisCache.setCacheObject(sendFreq,LocalDateTime.now(),resendTime,TimeUnit.MINUTES);
-
-        //发送验证码
-        String codeEncrypted;
-
-        try {
-            codeEncrypted = passwordEncoder.encode(mailService.sendVerificationCode(request.getEmail()));
-        } catch (MessagingException e) {
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Email sending failed");
-        }
-
-        request.setVerifyCode(codeEncrypted);
-
+        //发邮件
+        sendCode(request.getEmail());
 
         //先对密码加密
         request.setPassword(passwordEncoder.encode(request.getPassword()));
         request.setPassword(passwordEncoder.encode(request.getPassword()));
-        //存入redis
-        redisCache.setCacheObject("register:" + request.getEmail(),request,emailExpireTime,TimeUnit.MINUTES);
+        //基本信息存入redis
+        redisCache.setCacheObject("register:" + request.getEmail(),request,basicInfoExpireTime,TimeUnit.MINUTES);
 
-        return new GeneralResponseResult(HttpStatus.OK.value(),"The verification code has been sent successfully.");
+        return new GeneralResponseResult(ErrorCode.SUCCESS);
     }
+
+    @Override
+    public GeneralResponseResult sendCode(String email) {
+
+        //控制发送的频率,一分钟发一次
+        String sendFreq = "emailSendFreq:" + email;
+        System.out.println("5");
+        if(!Objects.isNull(redisCache.getCacheObject(sendFreq))){
+            throw new BusinessException(ErrorCode.TOO_MANY_REQUESTS);
+        }
+        redisCache.setCacheObject(sendFreq,LocalDateTime.now(),resendTime,TimeUnit.MINUTES);
+
+        //验证码密文
+        String code;
+        try{
+            code = mailService.sendVerificationCode(email);
+        } catch (MessagingException e) {
+            throw new SystemException(ErrorCode.EMAIL_SERVICE_ERROR);
+        }
+
+
+        //验证码存入redis
+        redisCache.setCacheObject("verifyCode:" + email,code,emailExpireTime,TimeUnit.MINUTES);
+        return new GeneralResponseResult(ErrorCode.SUCCESS);
+    }
+
 
     //验证后注册用户
     @Override
     public GeneralResponseResult verifyRegister(String email,String verifyCode) {
 
-         UserRegisterRequest request= redisCache.getCacheObject("register:" + email);
+        UserRegisterRequest request= redisCache.getCacheObject("register:" + email);
         if(Objects.isNull(request)){
-            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"Verification code error");
+            throw new AuthException(ErrorCode.USER_INFO_EXPIRED);
         }
+        //校验验证码是否正确
+        String code = redisCache.getCacheObject("verifyCode:" + email);
+        if(Objects.isNull(code) || code.isEmpty()){
+            throw new AuthException(ErrorCode.EMAIL_VERIFICATION_FAILED);
+        }
+        if(!code.equals(verifyCode)){
+            System.out.println(code);
+            throw new AuthException(ErrorCode.EMAIL_VERIFICATION_FAILED);
+        }
+
         //让验证码失效
         redisCache.deleteObject(email);
 
@@ -218,10 +214,10 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
         try{
             this.baseMapper.insert(user);
         } catch (DuplicateKeyException exception) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Email already registered");
+            throw new AuthException(ErrorCode.DB_OPERATION_ERROR.getCode(), "Email already registered");
         }
 
-        return new GeneralResponseResult(HttpStatus.OK.value(),"Successful registration");
+        return new GeneralResponseResult(ErrorCode.SUCCESS);
     }
